+++ components/tikv_util/src/yatp_pool/metrics.rs
@@ -2,8 +2,32 @@
 
 use lazy_static::lazy_static;
 use prometheus::*;
+use prometheus_static_metric::*;
+
+make_auto_flush_static_metric! {
+    pub label_enum PoolName {
+        unified_read_pool,
+        unknown,
+    }
+
+    pub label_enum ResourcePriority {
+        high,
+        medium,
+        low,
+        unknown,
+    }
+
+    pub struct YATPPoolScheduleWaitDurationVec: LocalHistogram {
+        "name" => PoolName,
+        "priority" => ResourcePriority,
+    }
+}
 
 lazy_static! {
+    pub static ref YATP_POOL_SCHEDULE_WAIT_DURATION_STATIC: YATPPoolScheduleWaitDurationVec =
+    auto_flush_from!(YATP_POOL_SCHEDULE_WAIT_DURATION_VEC, YATPPoolScheduleWaitDurationVec);
+
+
     pub static ref FUTUREPOOL_RUNNING_TASK_VEC: IntGaugeVec = register_int_gauge_vec!(
         "tikv_futurepool_pending_task_total",
         "Current future_pool pending + running tasks.",
@@ -19,8 +43,32 @@ lazy_static! {
     pub static ref YATP_POOL_SCHEDULE_WAIT_DURATION_VEC: HistogramVec = register_histogram_vec!(
         "tikv_yatp_pool_schedule_wait_duration",
         "Histogram of yatp pool schedule wait duration.",
-        &["name"],
+        &["name","priority"],
         exponential_buckets(1e-5, 4.0, 12).unwrap() // 10us ~ 41s
     )
     .unwrap();
+}
+
+
+impl From<u32> for ResourcePriority {
+    fn from(priority: u32) -> Self {
+        // the mapping definition of priority in TIDB repo,
+        // see: https://github.com/bufferflies/tidb/blob/8b151114546d6a02d8250787a2a3213620e30524/parser/parser.y#L1740-L1752
+        match priority {
+            1 => ResourcePriority::low,
+            8 => ResourcePriority::medium,
+            16 => ResourcePriority::high,
+            _ => ResourcePriority::unknown,
+        }
+    }
+}
+
+
+impl From<String> for PoolName {
+    fn from(name: String) -> Self {
+        match &name {
+            "unified_read_pool" => PoolName::unified_read_pool,
+            _ => PoolName::unknown,
+        }
+    }
 }
